# Conflicts:
#	src/main/java/com/trackit/enterprise/EnterpriseServiceImpl.java

+++ src/main/java/com/trackit/enterprise/EnterpriseServiceImpl.java
@@ -9,11 +9,10 @@ import com.trackit.driver.Driver;
 import com.trackit.driver.DriverRepo;
 import com.trackit.exception.EnterpriseAlreadyExistsException;
 import com.trackit.exception.EnterpriseNotFoundException;
-import com.trackit.utils.EnterpriseMappers;
+import static com.trackit.utils.EnterpriseMappers.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import static com.trackit.utils.EnterpriseMappers.mapToEnterpriseDTOList;
 
 @Service
 public class EnterpriseServiceImpl implements EnterpriseService{
@@ -36,28 +35,30 @@ public class EnterpriseServiceImpl implements EnterpriseService{
 	public EnterpriseDTO findEnterpriseById(Integer id) throws EnterpriseNotFoundException {
 		if(!enterpriseRepo.existsById(id))
 			throw new EnterpriseNotFoundException("No Enterprise Found with the Id ",id);
-		return EnterpriseMappers.mapToEnterpriseDTO(enterpriseRepo.getOne(id));
+		return mapToEnterpriseDTO(enterpriseRepo.getOne(id));
 	}
 
 
 
 	@Override
 	public EnterpriseDTO addEnterprise(EnterpriseDTO enterpriseDTO) throws EnterpriseAlreadyExistsException {
-		Enterprise enterprise = enterpriseRepo.saveAndFlush(EnterpriseMappers.maptoEnterprise(enterpriseDTO,carRepo, driverRepo));
-		return EnterpriseMappers.mapToEnterpriseDTO(enterprise);
+		Enterprise enterprise = enterpriseRepo.save(maptoEnterprise(enterpriseDTO,carRepo, driverRepo));
+		return mapToEnterpriseDTO(enterprise);
 	}
 
 	@Override
 	public EnterpriseDTO updateEnterprise(EnterpriseDTO enterpriseDTO) throws EnterpriseNotFoundException {
 		if (!enterpriseRepo.existsById(enterpriseDTO.getId()))
 			throw new EnterpriseNotFoundException("No Enterprise Found with the Id ",enterpriseDTO.getId());
-		Enterprise enterprise = enterpriseRepo.saveAndFlush(EnterpriseMappers.maptoEnterprise(enterpriseDTO,carRepo, driverRepo));
-		return  EnterpriseMappers.mapToEnterpriseDTO(enterprise);
+		Enterprise enterprise = enterpriseRepo.save(maptoEnterprise(enterpriseDTO,carRepo, driverRepo));
+		return  mapToEnterpriseDTO(enterprise);
 	}
 
 
 	@Override
-	public void deleteEnterprise(Integer id) {
+	public void deleteEnterprise(Integer id) throws EnterpriseNotFoundException {
+		if (!enterpriseRepo.existsById(id))
+			throw new EnterpriseNotFoundException("No Enterprise Found with the Id ",id);
 		enterpriseRepo.deleteById(id);
 		
 	}
